requirements.txt
alembic==1.14.0
blinker==1.9.0
click==8.1.8
flake8==7.1.1
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
load-dotenv==0.1.0
Mako==1.3.8
MarkupSafe==3.0.2
mccabe==0.7.0
packaging==24.2
pluggy==1.5.0
pycodestyle==2.12.1
pyflakes==3.2.0
pytest==8.3.4
pytest-flask==1.3.0
python-dotenv==1.0.1
SQLAlchemy==2.0.37
typing_extensions==4.12.2
Werkzeug==3.1.3
WTForms==3.2.1
